fix type char/uint8_t for text

+++ src/hal/mcu/atmega2560/hal_gpio.c
@@ -45,16 +45,25 @@ static const gpio_mcu_port_t gpioMcuPorts[GPIO_PORT_COUNT] = {
 void hal_gpioInitPin(const gpio_pin_item_t *pin)
 {
 
-	if( pin->mode == GPIO_PIN_MODE_INPUT ) { *(gpioMcuPorts[pin->port_index].ddr) &= (uint8_t)~(1 << pin->number); }
-	if( pin->mode == GPIO_PIN_MODE_OUTPUT_PP ) { *(gpioMcuPorts[pin->port_index].ddr) |= (1 << pin->number); }
+	if( pin->mode == GPIO_PIN_MODE_INPUT )
+	{
+		*(gpioMcuPorts[pin->port_index].ddr) &= (uint8_t)~(1u << pin->number);
+	}
+	if( pin->mode == GPIO_PIN_MODE_OUTPUT_PP )
+	{
+		*(gpioMcuPorts[pin->port_index].ddr) |= (uint8_t)(1u << pin->number);
+	}
 
-	if( pin->pull == GPIO_PIN_PULL_UP ) { *(gpioMcuPorts[pin->port_index].port) |= (1 << pin->number); }
+	if( pin->pull == GPIO_PIN_PULL_UP )
+	{
+		*(gpioMcuPorts[pin->port_index].port) |= (uint8_t)(1u << pin->number);
+	}
 }
 
 void hal_gpioWritePin(const gpio_pin_item_t *pin, bool value)
 {
-	if( value ) { *(gpioMcuPorts[pin->port_index].port) |= (1 << pin->number); }
-	else { *(gpioMcuPorts[pin->port_index].port) &= (uint8_t)~(1 << pin->number); }
+	if( value ) { *(gpioMcuPorts[pin->port_index].port) |= (uint8_t)(1u << pin->number); }
+	else { *(gpioMcuPorts[pin->port_index].port) &= (uint8_t)~(1u << pin->number); }
 }
 
 bool hal_gpioReadPin(const gpio_pin_item_t *pin)
